+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@
 
 ### What's New
 - Add document with rationale behind analysis scope (CPU architectures, FPU instructions, metrics, ...) & with rigorous references behind obtained data. 
+- Add support for SSE2- and ARM-specific FPU instruction latencies
 
 ### Improvements
 /